requirements.txt
streamlit
reportlab
pandas
pillow